Add UnitCurveCompatibilityTests
