+++ Tests/OpenSwiftUICoreTests/Animation/Animation/UnitCurveTests.swift
@@ -5,213 +5,9 @@
 @testable import OpenSwiftUICore
 import Testing
 
-// MARK: - UnitCurveTests [Implmeneted by Copilot]
+// MARK: - UnitCurveTests
 
 struct UnitCurveTests {
-    // MARK: - Static Curve Properties
-
-    @Test
-    func linearCurve() {
-        let curve = UnitCurve.linear
-
-        #expect(curve.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(curve.value(at: 0.5).isApproximatelyEqual(to: 0.5))
-        #expect(curve.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-
-        #expect(curve.velocity(at: 0.0).isApproximatelyEqual(to: 1.0))
-        #expect(curve.velocity(at: 0.5).isApproximatelyEqual(to: 1.0))
-        #expect(curve.velocity(at: 1.0).isApproximatelyEqual(to: 1.0))
-    }
-
-    @Test
-    func easeInCurve() {
-        let curve = UnitCurve.easeIn
-
-        #expect(curve.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(curve.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-
-        let midValue = curve.value(at: 0.5)
-        #expect(midValue > 0.0)
-        #expect(midValue < 0.5)
-    }
-
-    @Test
-    func easeOutCurve() {
-        let curve = UnitCurve.easeOut
-
-        #expect(curve.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(curve.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-
-        let midValue = curve.value(at: 0.5)
-        #expect(midValue > 0.5)
-        #expect(midValue < 1.0)
-    }
-
-    @Test
-    func easeInOutCurve() {
-        let curve = UnitCurve.easeInOut
-
-        #expect(curve.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(curve.value(at: 0.5).isApproximatelyEqual(to: 0.5))
-        #expect(curve.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-
-        let quarterValue = curve.value(at: 0.25)
-        #expect(quarterValue > 0.0)
-        #expect(quarterValue < 0.25)
-
-        let threeQuarterValue = curve.value(at: 0.75)
-        #expect(threeQuarterValue > 0.75)
-        #expect(threeQuarterValue < 1.0)
-    }
-
-    @Test
-    func easeInEaseOutDeprecated() {
-        let curve = UnitCurve.easeInEaseOut
-        let easeInOut = UnitCurve.easeInOut
-
-        #expect(curve.value(at: 0.0).isApproximatelyEqual(to: easeInOut.value(at: 0.0)))
-        #expect(curve.value(at: 0.5).isApproximatelyEqual(to: easeInOut.value(at: 0.5)))
-        #expect(curve.value(at: 1.0).isApproximatelyEqual(to: easeInOut.value(at: 1.0)))
-    }
-
-    @Test
-    func circularEaseInCurve() {
-        let curve = UnitCurve.circularEaseIn
-
-        #expect(curve.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(curve.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-
-        let midValue = curve.value(at: 0.5)
-        #expect(midValue > 0.0)
-        #expect(midValue < 0.5)
-    }
-
-    @Test
-    func circularEaseOutCurve() {
-        let curve = UnitCurve.circularEaseOut
-
-        #expect(curve.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(curve.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-
-        let midValue = curve.value(at: 0.5)
-        #expect(midValue > 0.5)
-        #expect(midValue < 1.0)
-    }
-
-    @Test
-    func circularEaseInOutCurve() {
-        let curve = UnitCurve.circularEaseInOut
-
-        #expect(curve.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(curve.value(at: 0.5).isApproximatelyEqual(to: 0.5))
-        #expect(curve.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-    }
-
-    // MARK: - Bezier Curve Creation
-
-    @Test
-    func bezierCurveCreation() {
-        let startPoint = UnitPoint(x: 0.25, y: 0.1)
-        let endPoint = UnitPoint(x: 0.75, y: 0.9)
-        let curve = UnitCurve.bezier(startControlPoint: startPoint, endControlPoint: endPoint)
-
-        #expect(curve.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(curve.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-    }
-
-    @Test
-    func bezierCurveWithExtremeControlPoints() {
-        let startPoint = UnitPoint(x: -0.5, y: 2.0)
-        let endPoint = UnitPoint(x: 1.5, y: -1.0)
-        let curve = UnitCurve.bezier(startControlPoint: startPoint, endControlPoint: endPoint)
-
-        #expect(curve.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(curve.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-    }
-
-    // MARK: - Value Function
-
-    @Test
-    func valueAtBoundaryConditions() {
-        let curve = UnitCurve.easeInOut
-
-        #expect(curve.value(at: -0.5).isApproximatelyEqual(to: 0.0))
-        #expect(curve.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(curve.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-        #expect(curve.value(at: 1.5).isApproximatelyEqual(to: 1.0))
-    }
-
-    @Test
-    func valueAtVariousProgressPoints() {
-        let curve = UnitCurve.linear
-
-        for progress in stride(from: 0.0, through: 1.0, by: 0.1) {
-            let value = curve.value(at: progress)
-            #expect(value >= 0.0)
-            #expect(value <= 1.0)
-        }
-    }
-
-    // MARK: - Velocity Function
-
-    @Test
-    func velocityAtBoundaryConditions() {
-        let curve = UnitCurve.linear
-
-        #expect(curve.velocity(at: -0.5).isApproximatelyEqual(to: 1.0))
-        #expect(curve.velocity(at: 0.0).isApproximatelyEqual(to: 1.0))
-        #expect(curve.velocity(at: 1.0).isApproximatelyEqual(to: 1.0))
-        #expect(curve.velocity(at: 1.5).isApproximatelyEqual(to: 1.0))
-    }
-
-    @Test
-    func velocityForCircularCurves() {
-        let easeIn = UnitCurve.circularEaseIn
-        let easeOut = UnitCurve.circularEaseOut
-        let easeInOut = UnitCurve.circularEaseInOut
-
-        #expect(easeIn.velocity(at: 0.5) > 0.0)
-        #expect(easeOut.velocity(at: 0.5) > 0.0)
-        #expect(easeInOut.velocity(at: 0.5) > 0.0)
-    }
-
-    // MARK: - Inverse Property
-
-    @Test
-    func linearCurveInverse() {
-        let curve = UnitCurve.linear
-        let inverse = curve.inverse
-
-        #expect(inverse.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(inverse.value(at: 0.5).isApproximatelyEqual(to: 0.5))
-        #expect(inverse.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-    }
-
-    @Test
-    func circularCurveInverse() {
-        let easeIn = UnitCurve.circularEaseIn
-        let easeOut = UnitCurve.circularEaseOut
-        let easeInOut = UnitCurve.circularEaseInOut
-
-        #expect(easeIn.inverse.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(easeIn.inverse.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-        #expect(easeOut.inverse.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(easeOut.inverse.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-        #expect(easeInOut.inverse.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(easeInOut.inverse.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-    }
-
-    @Test
-    func bezierCurveInverse() {
-        let startPoint = UnitPoint(x: 0.25, y: 0.1)
-        let endPoint = UnitPoint(x: 0.75, y: 0.9)
-        let curve = UnitCurve.bezier(startControlPoint: startPoint, endControlPoint: endPoint)
-        let inverse = curve.inverse
-
-        #expect(inverse.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(inverse.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-    }
-
     // MARK: - CubicSolver
 
     @Test
@@ -234,10 +30,7 @@ struct UnitCurveTests {
 
         #expect(solver.value(at: 0.0).isApproximatelyEqual(to: 0.0))
         #expect(solver.value(at: 1.0).isApproximatelyEqual(to: 1.0))
-
-        let midValue = solver.value(at: 0.5)
-        #expect(midValue >= 0.0)
-        #expect(midValue <= 1.0)
+        #expect(solver.value(at: 0.5).isApproximatelyEqual(to: 0.5))
     }
 
     @Test
@@ -245,31 +38,7 @@ struct UnitCurveTests {
         let startPoint = UnitPoint(x: 0.25, y: 0.1)
         let endPoint = UnitPoint(x: 0.75, y: 0.9)
         let solver = UnitCurve.CubicSolver(startControlPoint: startPoint, endControlPoint: endPoint)
-
-        let velocity = solver.velocity(at: 0.5)
-        #expect(velocity.isFinite)
-    }
-
-    // MARK: - Hashable and Sendable
-
-    @Test
-    func hashableConformance() {
-        let curve1 = UnitCurve.linear
-        let curve2 = UnitCurve.linear
-        let curve3 = UnitCurve.easeIn
-
-        #expect(curve1.hashValue == curve2.hashValue)
-        #expect(curve1.hashValue != curve3.hashValue)
-    }
-
-    @Test
-    func equatableConformance() {
-        let curve1 = UnitCurve.linear
-        let curve2 = UnitCurve.linear
-        let curve3 = UnitCurve.easeIn
-
-        #expect(curve1 == curve2)
-        #expect(curve1 != curve3)
+        #expect(solver.velocity(at: 0.5).isApproximatelyEqual(to: 0.399, absoluteTolerance: 0.001))
     }
 
     // MARK: - Edge Cases
@@ -280,8 +49,8 @@ struct UnitCurveTests {
 
         #expect(curve.value(at: -1000.0).isApproximatelyEqual(to: 0.0))
         #expect(curve.value(at: 1000.0).isApproximatelyEqual(to: 1.0))
-        #expect(curve.velocity(at: -1000.0) >= 0.0)
-        #expect(curve.velocity(at: 1000.0) >= 0.0)
+        #expect(curve.velocity(at: -1000.0).isApproximatelyEqual(to: 0.0))
+        #expect(curve.velocity(at: 1000.0).isApproximatelyEqual(to: 0.0))
     }
 
     @Test
@@ -290,8 +59,9 @@ struct UnitCurveTests {
         let curve = UnitCurve.bezier(startControlPoint: point, endControlPoint: point)
 
         #expect(curve.value(at: 0.0).isApproximatelyEqual(to: 0.0))
-        #expect(curve.value(at: 0.5).isApproximatelyEqual(to: 0.5))
         #expect(curve.value(at: 1.0).isApproximatelyEqual(to: 1.0))
+        #expect(curve.value(at: 0.5).isApproximatelyEqual(to: 0.5))
+        #expect(curve.velocity(at: 0.5).isApproximatelyEqual(to: 1.0))
     }
 
     @Test
